pyproject.toml
[project]
name = "wichtelapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-allauth>=65.13.1",
    "django-debug-toolbar>=6.1.0",
    "django-simple-deploy[fly-io]>=1.4.0",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.13",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "djhtml>=3.0.10",
    "ruff>=0.14.7",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".github",
    ".hg",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "migrations",
    "node_modules",
    "static",
]
# Same as Black.
line-length = 120
target-version = "py313"

[tool.ruff.lint]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F"]
ignore = ["E501", "E741"]                # temporary
per-file-ignores = {}
# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
unfixable = []
